refactor: move handler functions to their own directory
