+++ src/handlers/README.md
@@ -1,0 +1,3 @@
+# Handlers
+
+This directory contains functions that handle various types of events. Currently "events" refer to user interactions and online peer requests.
